pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cheeksbase"
version = "0.1.0"
description = "Agent-first data platform with YAML-only connectors"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "duckdb>=1.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
    "pydantic>=2.0",
    "fastmcp>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pyarrow>=14.0",
    "faker>=24.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
]

[project.scripts]
cheeksbase = "cheeksbase.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["cheeksbase"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B905",  # zip() without strict= — len(cols)==len(rows) is guaranteed by DB
]

[tool.ruff.lint.isort]
known-first-party = ["cheeksbase"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true